Problème réglé pour l'affichage des chunks 32x32x32 (c'était un problème de type pour les indices, le type short ne suffisait pas, il fallait des int

+++ include/Voxel.hpp
@@ -6,7 +6,7 @@ class Voxel {
     private:
         glm::vec3 backBottomLeftCorner;
         std::vector<glm::vec3> vertices;
-        std::vector<unsigned short> indices;
+        std::vector<unsigned int> indices;
         //int faceId;
     public:
         Voxel(glm::vec3 position, int numBlock);
@@ -14,5 +14,5 @@ class Voxel {
         glm::vec3 getPoint();
         void updatePoint(glm::vec3 motion);
         std::vector<glm::vec3> getVertices();
-        std::vector<unsigned short> getIndices();
+        std::vector<unsigned int> getIndices();
 };
